+++ python/fast_mlsirm/__init__.py
@@ -18,7 +18,7 @@ from .inference import oakes_standard_errors as oakes_standard_errors, observed_
 from .linking import link_fixed_item_parameters as link_fixed_item_parameters
 from .linking import irt_link as irt_link, IrtLinkResult as IrtLinkResult
 from .equating import equate_observed_scores as equate_observed_scores, equate_neat as equate_neat, EquateResult as EquateResult, equate_observed_scores_kernel as equate_observed_scores_kernel, loglinear_smooth as loglinear_smooth, equate_neat_linear as equate_neat_linear, equating_standard_errors as equating_standard_errors
-from .rt import fit_response_times as fit_response_times, RtFit as RtFit
+from .rt import fit_response_times as fit_response_times, RtFit as RtFit, fit_speed_accuracy as fit_speed_accuracy
 from .report import render_diagnostics_report as render_diagnostics_report
 from .validation import (ValidationVerdict as ValidationVerdict,
                          validate_judge as validate_judge)
@@ -81,6 +81,7 @@ __all__ = [
     "equating_standard_errors",
     "fit_response_times",
     "RtFit",
+    "fit_speed_accuracy",
     "export_serving_bundle",
     "fit",
     "fit_polytomous",